[ffx][daemon] Use build-dir for socket in-tree
Instead of using the global socket, use a per-build-dir daemon socket
path when using ffx in-tree. This prevents frequent fights over the
daemon socket when jumping between in and out of tree use cases, as well
as between different build outputs.

Also fixed a bunch of error paths involved to be clearer and/or do more
sensible things and not present as BUG: when they happen.

Fixed: 109129
Change-Id: I1e4220660bc52339c8e4c23a13dcc4560dad0b6e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/723317
Commit-Queue: Megan Batty <mgnb@google.com>
Reviewed-by: Shai Barack <shayba@google.com>