[tools][fuzz] Add test coverage for non-integer time flags
While flags such as max_total_time are technically defined as an integer
flag in libFuzzer, it uses strtol() which drops anything after the first
non-numeric character without complaining. It turns out that ClusterFuzz
sometimes passes floats but because of this leniency on the part of
libFuzzer it has not caused any issues.

The corresponding `ffx fuzz` option, however, previously explicitly
rejected float values, so this expands the undercoat e2e tests to
prevent regression on this behavior.

Bug: 108878
Change-Id: Icd1686405519c9afe59b98d7cfe14891f0d36804
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/723530
Reviewed-by: Cameron Finucane <eep@google.com>
Commit-Queue: Cameron Finucane <eep@google.com>
Reviewed-by: Aaron Green <aarongreen@google.com>