[kernel][vm] Remove lock from content size field
This change removes the lock from the content size field in favor of an
atomic. This is in preparation for a following change that allows for
content-expanding writes to mutate the content size as progress is made
without needing to hold the `ContentSizeManager`'s lock.

As mentioned in code comments, removing this lock is logically sound
because the queues in `ContentSizeManager` uphold the invariant that
only one content size mutating operation can occur at any given time.

Bug: 108771
Bug: 107119
Change-Id: I80100d72f3ab00313a23ec6d2c0a88217cc79289
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/724649
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>
Commit-Queue: Chris Tam <godtamit@google.com>
Reviewed-by: Adrian Danis <adanis@google.com>