[driver_host] Fix power state map indexing
An off-by-1 error for indexing into the system power state map seems to
have lead to certain code paths being optimized away. Specifically in
LTO builds, the `DEVICE_SUSPEND_FLAG_REBOOT_KERNEL_INITIATED` switch
case in `device_get_dev_power_state_from_mapping` was not being executed
as expected. Presumably this switch path was optimized away because the
result it produced is only used to index into an array, which would be
invalid in this case.

Since the SystemPowerState is a FIDL enum which is a 1-based index, we
should actually be using `sys_state - 1` as the index into the system
power state map.

Ultimately, it doesn't seem like the value from the map is used
anywhere, which is probably why this went undetected for so long.

Test: fx test driver-host-tests
Test: fx test host_x64/obj/src/tests/oom/oom_tests.sh
Fixed: 105795
Change-Id: I96ced09aa8af01f1d0a27e6b229d358ff0bb9172
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/723481
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Fuchsia-Auto-Submit: Patrick Shickel <pshickel@google.com>
Commit-Queue: Patrick Shickel <pshickel@google.com>