[fidl][rust] Sync Event Handling TXID Check
Modify FIDL Rust sync event handling to check the transaction IDs.
Previously, sync event handling would not check transaction IDs. This
would lead to a scenario where a reply to a two-way method could turn up
as an event.

For example, if a method is called in the sync bindings but the deadline
expires before the response is received, the message will be pulled out
by the event handling instead. Without unknown interaction handling this
would result in an UnknownOrdinal message, but with unknown interaction
handling, if the method was flexible, the method would turn up as an
unknown flexible event, even though its actually a reply to an expired
two-way method call.

This behavior is incorrect. A two-way method reply should never appear
as an event, and other binding have the correct behavior. Async event
handling (should) also have the correct behavior because it only inserts
messages into the event queue if they have TXID zero, but this will be
tested later when the Rust async client bindings are added to the
dynsuite.

This difference was discovered as part of adding the Rust bindings to
the dynsuite client tests, so the test for this bug is in the dynsuite.

Bug: 88366
Change-Id: I09b2c02e5bea1cb70149de24eaedc3fe68388362
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/724748
Reviewed-by: Mitchell Kember <mkember@google.com>
Commit-Queue: Zachary Stewart <zstewart@google.com>