[ffx] Extend `AddTarget` to return more info.
This extends `AddTarget` to return more info about errors. For example,
there may be an ABI mismatch, but it is important that the user know
what the ABI mismatch is (the local vs the remote ABI). This sends a
struct to the user when there is an error instead of extending an error
struct. Without this change, if there is an error that has additional information
included with it (like a specific ABI compatibility), the added target will be
automatically removed if requested to do so in the config.

Current behavior should be unchanged, as the return log of errors is not
intended to be parsed, simply relayed to the user.

Fixed: 107759

Change-Id: Ieff7c2c23807e89442aba847b8f6d7b45a654e25
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/722124
API-Review: Dave Schuyler <dschuyler@google.com>
Reviewed-by: Megan Batty <mgnb@google.com>
Commit-Queue: Andrew Davies <awdavies@google.com>