[ffx][fuzz] Stop does not need to connect
This CL adjusts `ffx fuzz stop` to avoid trying to first execute an
`ffx fuzz attach` command. The latter results in a
`fuchsia.fuzzer.Manager/Connect` FIDL request, which is pointless when
the former is trying to send a `fuchsia.fuzzer.Manager/Stop` request.

Separating these requires that the `StopSubcommand` not be derived from
a `StopShellCommand`, and provide its own optional URL as an argument.
If the URL is not provided, `stop` will still use the one from a
previous interactive `attach` command, if present.

Change-Id: I76515d41d31452ccb6f7e9a4b1526c1520fa8b37
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/723494
Commit-Queue: Aaron Green <aarongreen@google.com>
Reviewed-by: Cameron Finucane <eep@google.com>