[ffx][coverage] Add a few flags
--path-remapping: rewrite source file paths when exporting coverage
--compilation-dir: base dir for relative coverage mapping paths

These flags combined provides the ability to correctly set paths of
source file in the final coverage reports.

Note no default values are provided for these flags, because they can
be quite different depending on the environment ffx is executed in. It
is up to the corresponding wrappers (e.g. `fx coverage` for in-tree,
`bazel coverage` for OOT) to correctly pass in correct values for these
flags when necessary.

Bug: 103896
Change-Id: I16365911cf769e9bc403454a63a65daab7f14b0c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/720509
Commit-Queue: Jay Zhuang <jayzhuang@google.com>
Reviewed-by: Petr Hosek <phosek@google.com>
Reviewed-by: Satsuki Ueno <satsukiu@google.com>