[iwlwifi] pass all multicast packets to driver
Currently we only ask the firmware to pass 3 fixed multicast
addresses to the driver. However, it is not enough for IPv6.

So the new code asks the firmware to pass all multicast addresses
to the driver so that the driver can forward the IPv6 multicast
packets to the netstack.

Bug: 108681, 100129, 99200

Change-Id: I8b708cab35030a23696740feb8c4eba1da9c88ca
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/724077
Reviewed-by: John Tan <johntan@google.com>
Commit-Queue: Louis Lo <yjlou@google.com>