[magma] Verify imported and exported symbols from magma_client
magma_client needs to import only symbols on the magma allowlist so that
OOT-built ICDs will also be able to abide by the allowlist. We also
want to verify the symbols it exports to reduce the risk of undetected
API changes, and ensure clients don't accidentally use symbols from it
that we aren't expecting.

The existing tools to verify symbols use a shared library, so the static
library is first compiled into a shared library before doing the checks.

trace_generate_nonce was accidentally being exported, so localize that
symbol.

Change-Id: I74290f80068b0d06164b96dabdf76c080bbe6753
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/716516
Reviewed-by: Craig Stout <cstout@google.com>
Commit-Queue: John Bauman <jbauman@google.com>