[build] Remove export-compile-commands switch
The toplevel .gn file now globally exports a compilation database for
the //:default target. It is no longer necessary to pass a switch to GN
to make this happen, and the uses of this switch have been removed.

Remove the zircon_comdb_filter variable. This is never used.

Bug: 109452

Change-Id: I0f042d83f10772d9f648ae5076af53d133f921fd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/725327
Reviewed-by: Adam Barth <abarth@google.com>
Commit-Queue: Brett Wilson <brettw@google.com>
Reviewed-by: Oliver Newman <olivernewman@google.com>