[wlan][policy] add logs for removing saved network
Add logging for the case where an API call comes in to remove a saved
network and the SSID matches but the security type doesn't match. Log
the security type of the network that matches. This will help debugging
cases where the network is not correctly specified to be removed.

This change also gets rid of the key from the map of saved networks if
the only saved network is removed.

Test: manually tested logs
Bug: 105235
Change-Id: Iea405be1d3e9455bc2eae5423282755c450dc027
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/678945
Reviewed-by: Josh Armour <jarmour@google.com>
Reviewed-by: Kevin Sakuma <sakuma@google.com>
Reviewed-by: Marc Khouri <mnck@google.com>
Commit-Queue: Naomi Mccracken <nmccracken@google.com>