[testgen][rust] Rename lib.rs -> testgen.rs
lib.rs is a module name that is the entry point for rust libraries, and
it has become a warning to name a module 'lib' in new toolchains.

Rename the generated module to 'testgen' which makes it clear where the
code came from and hints that it will be (eventually) refreshed.

Bug: 109622
Test: refactor; fx build && CQ
Change-Id: If50b653324a9711154d17cc3003779a913eaf6b0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/726543
Fuchsia-Auto-Submit: Marie Janssen <jamuraa@google.com>
Commit-Queue: Marie Janssen <jamuraa@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Yuan Zhi <yuanzhi@google.com>