[virtualization] Add integration test for container startup.
Add a new type of enclosed guest: 'TerminaContainerEnclosedGuest'. This
extends the existing TerminaEnclosedGuest to complete the container
startup. Container images are pulled from a local file on a virtio-block
device instead of the internet.

In terms of performance, running the container startup takes ~1 minute
to complete so test that use the Termina container will be implicitly
longer. For now we just have a simple test to verify that the container
has successfully been created.

Bug: 109706
Multiply: termina_container_tests
Change-Id: I9c0891cbfe0fcc9e9b1c0d29d211ba0025cd63d9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/720351
Commit-Queue: Tim Detwiler <tjdetwiler@google.com>
Reviewed-by: Jonathan Bailey <jonathanbailey@google.com>