[tools][fuzz] Allow stop_instance to run with older handles
If a handle is from an older version that had fewer fields, we can still
do a best-effort job of cleaning up instead of declaring the handle
invalid.

This scenario occurs when a ClusterFuzz build with a newer version of
undercoat needs to clean up after a previously-running build that had an
earlier version of undercoat with a different handle format. This won't
work for all possible version mismatches, but is meant to be a simple
stop-gap solution.

Change-Id: If9c7cfe25b9d61cd945976683d3edea1b35a69ba
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/724823
Reviewed-by: Aaron Green <aarongreen@google.com>
Commit-Queue: Cameron Finucane <eep@google.com>