[check-licenses] Move project filtering to cmd package.
Currently, check-licenses calls the gn executable at the end of the
execution, in order to filter out the projects that aren't in the
current build.

This works, but makes it difficult to change the behavior of the tool in
some situations. E.g. in the future, if we know that a given project
hasn't changed from the previous run, we could potentially skip
processing that project entirely. But we need to do that analysis up
front.

This CL moves all of the filtering logic to the cmd package. In the
future, the cmd package will also do some preprocessing to recognize
areas of the tree that haven't changed recently, and skip processing
those parts of the tree.

Change-Id: I51de5856934ce90909ce16048715fca98f6bcf21
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/726502
Fuchsia-Auto-Submit: Jerry Belton <jcecil@google.com>
Reviewed-by: Yuxiao Sun <yuxiaosun@google.com>
Commit-Queue: Jerry Belton <jcecil@google.com>