[reland][virtualization] Add integration test for container startup.
When this first landed it caused the sherlock-debug builders to run out
of space. Since we don't run this test on ARM we just need to make sure
we don't include the container images on ARM either to avoid this
problem.

Add a new type of enclosed guest: 'TerminaContainerEnclosedGuest'. This
extends the existing TerminaEnclosedGuest to complete the container
startup. Container images are pulled from a local file on a virtio-block
device instead of the internet.

In terms of performance, running the container startup takes ~1 minute
to complete so test that use the Termina container will be implicitly
longer. For now we just have a simple test to verify that the container
has successfully been created.

Bug: 109706
Multiply: termina_container_tests
Change-Id: Ia1f09be35b4d3311e0a45738a1874cbd7ee5d10d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/727317
Commit-Queue: Tim Detwiler <tjdetwiler@google.com>
Reviewed-by: David Hastings <dahastin@google.com>