[gazelle] Tidy up the gazelle prototype.
This splits the previous `WindowManager` into two layers: `View` and
`Manager`. The `View` doesn't do anything smart at this point: it just
interacts with Flatland directly.

The `Manager` has policy in it (e.g., as soon as the child for a new
window is attached, transfer focus to it). It also handles queuing up
future work and executing it in the background.

This doesn't change any functionality, so it's still pretty much a
toy, but I think (hope) this'll be nicer to work with.

Change-Id: Ib6734bf26e6755932db82b7d65ec2b56b9b03c2e
Reviewed-on: https://fuchsia-review.googlesource.com/c/experiences/+/727025
Reviewed-by: Sanjay Chouksey <sanjayc@google.com>
Commit-Queue: Hunter Freyer <hjfreyer@google.com>