[vaapi] Don't always reconfig output buffers
When a mid-stream configuration is requested by the decoder check to see
if the current output buffers are large enough to hold the new size. If
so there isn't a need to reallocate buffers from sysmem. Instead just
send the new constraints to the client and keep decoding. Also added
some checks so that we make sure the new size is supported by the
underlying hardware so we don't get a nasty libva error.

Fixed: 102737

Change-Id: I5e7a50604a63b499664c1fe07b8a53f5b7499d15
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/696702
Reviewed-by: Dustin Green <dustingreen@google.com>
Commit-Queue: Stefan Bossbaly <stefanbossbaly@google.com>