[starnix] Serve view backed by framebuffer
This change makes Starnix serve a view backed by a framebuffer for
components that specify the `framebuffer` feature.

The framebuffer is always created and stored on the kernel, but is only
exposed in galaxies with the `framebuffer` feature.

Change-Id: I43c680741caf926f1a85e62b0f6b7e7de7abce24
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/727708
Fuchsia-Auto-Submit: Kevin Lindkvist <lindkvist@google.com>
Reviewed-by: Theodore Dubois <tbodt@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>