[ui][input-pipeline] move "recognized" messages to inspect
Move gesture start messages to inspect. Along the way,
add details on latency from the first event of a gesture,
until the gesture was recognized.

The latency is reported in two forms:
1. The number of touchpad events from the first event until
   the gesture was recognized.
2. The elapsed monotonic time between the driver timestamp of the
   first event, and the monotonic time when the gesture was
   recognized.

The first measure is intended to help diagnose issues with
our algorithms. For example:
* A recognizer takes too long to reject a gesture, slowing
  down the recognizer that eventually claims the gesture.
* The thresholds are too high, causing, e.g. motion recognition
  to take longer than it should.

The second measure is intended to diagnose slowness due to
system load. E.g., if the latency in the number of events
is 4, but the latency in elapsed monotonic milliseconds is
100, then the recognition was likely slowed by system load.
(The input reports should arrive from the touchpad to the
host CPU at 125 Hz, or 8 milliseconds apart.)

Bug: 107948
Test: fx test input_pipeline_lib_tests
Change-Id: I6eeb2ea17c2ea889b5a717a2165e8b28776e64c3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/726266
Fuchsia-Auto-Submit: Mukesh Agrawal <quiche@google.com>
Reviewed-by: Jianpeng Chao <chaopeng@google.com>
Reviewed-by: Mukesh Agrawal <quiche@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>