[starnix] Don't trash mappings in mprotect
Previously mprotect would replace the range given with an expanded
version of the first mapping in the range. This is wrong if there is
more than one thing mapped in the range.

Fixes weird issues in art, which turned out to be a result of calls to
mprotect that covered multiple mappings.

Change-Id: Ib2ed3e901b048d90eac23185dc2a365ab7982001
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/727189
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Commit-Queue: Theodore Dubois <tbodt@google.com>