[kernel][mmu] Extra checks and logs for arm64 PTs
Expands the check in `first_used_page_table_entry` to attempt to
validate that the entry found is actually a valid entry, instead of
just not an exact invalid entry. This aims to check for the scenario
where an entry has been corrupted and so it fails a full comparison
with zero, but it doesn't actually have a valid entry.

For the Destroy path an additional method `DebugFindFirstLeafMapping`
is added to attempt to find and display the page table entry, valid or
corrupted, that is holding page tables allocated.

Bug: 79118
Change-Id: Id22d7e308ac2c2c1b30a2e618bc14feca9612857
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/727582
Commit-Queue: Adrian Danis <adanis@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Reviewed-by: Travis Geiselbrecht <travisg@google.com>