[netstack3] Move try_update_addr to entry API
This allows doing fun stuff like returning a new Entry to the moved
value on success, which can then be used for further manipulation. This
will be used to update the IP socket of connected UDP sockets when the
socket is bound to or unbound from a device.

Bug: 108663
Change-Id: Ibf71a71f6800ac02ab74536886a339a65e50fc2d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/726466
Commit-Queue: Alex Konradi <akonradi@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Reviewed-by: Zeling Feng <zeling@google.com>