[dfv2] Fix node Remove lifetime
There is a flake where the node is freed halfway through
the Remove function.

The callstack looks like:
- Someone calls Remove on the parent Node.
  - The parent Node calls remove on its children.
    - The last child see's that its parent doesn't have any children
      so it calls Remove back on the parent.
      - The parent's new Remove function frees it.
    - The child is freed.
  - The parent continues the Remove function but it's been freed
    (segfault).

This is a simple fix, all we have to do is keep a shared pointer
alive during Remove to make sure we are never freed in the process
of calling this function.

Fixed: 109777

Change-Id: I4eb196376c2720a986544fbeeae5bc9f93448585
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/727030
Commit-Queue: David Gilhooley <dgilhooley@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>