[utest] Set fake UTC backstops to zero.
These tests use ZX_CLOCK_OPT_AUTOSTART, which creates a clock that is a
clone of the monotonic clock. What is less clearly documented is that
the backstop time provided must be in the past for the monotonic clock.
The upshot of this is that these tests currently rely on the fact that
they were run at least 123,456,789ns after the host boots. As it
happens, we spent a not insiginficant amount of boot time allocating an
aspace containing a single 16GiB mapping for EFI. A future CL will
change this, which causes these tests to start flaking (or failing
outright).

This CL reduces the backstop to 0 so that the tests pass regardless of
the time the kernel takes to boot.

Bug: 110001
Change-Id: If627b052a7ed420685d036fb9183b917fede8bd4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/726882
Reviewed-by: John Grossman <johngro@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Simon Shields <simonshields@google.com>