[fxfs] Implement syncfs
syncfs(2) is implemented in fdio as a call to sync() on the passed
directory fd.

Fxfs currently does nothing for sync() on a Directory; fix that to
execute a filesystem sync.

For consistency with minfs, we also flush the device, although it's
debatable if we need that.

Bug: 96085
Change-Id: Ic111fbca75968ad60afaa1d679c0ceff38dead6d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/728323
Reviewed-by: Martin Lindsay <mlindsay@google.com>
Commit-Queue: James Sullivan <jfsulliv@google.com>