[build][ctf] Remove deps from verify_deps_in_sdk
This CL changes verify_deps_in_sdk, a template that verifies deps, so
that the verification target it produces does not itself depend on the
deps. To clarify this, renamed the argument from deps to deps_to_verify.

The verification checking only depends on information produced at gn gen
time, so it's unnecessary to introduce edges in the ninja graph. This
was also causing problems for https://fxbug.dev/108354 because the
verification target was getting expanded in the default toolchain, but
the deps did not exist in the default toolchain.

Bug: 108354 (enforce _hlcpp suffixes)
Bug: 87864 (unwanted dependencies in CTF templates)
Change-Id: Id96adb570e9779e8ccd8ff6b439c9629229c01a6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/728085
Reviewed-by: John Shamoon <johnshamoon@google.com>
Reviewed-by: Kendal Harland <kjharland@google.com>
Commit-Queue: Mitchell Kember <mkember@google.com>