[netstack3] Update conn IpSock address with device
When SO_BINDTODEVICE is set for a connected UDP socket, attempt to
update the device as well. If the result is unroutable, reject the
attempt to set the device. This is different than Linux behavior because
it makes setting SO_BINDTODEVICE fail in a case that Linux would
allow.

Note that this needs to be documented per fxbug.dev/109358.

Bug: 108663, 109358
Change-Id: Ib0c14bda19ffc16e2c99f960d460177b01b1d853
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/726468
Reviewed-by: Ghanan Gowripalan <ghanan@google.com>
Commit-Queue: Alex Konradi <akonradi@google.com>
Fuchsia-Auto-Submit: Alex Konradi <akonradi@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>