[tools/fidl] Fix strictness bug when diffing APIs
After this change ElementStr.HasStrictness returns true iff the underlying FIDL type has a notion of strictness as determined by
the FIDL language spec. Before this change it only considered a
type as 'having strictness' if the encoded JSON contained a
'strictness' property.

This also fixes a bug where struct/members added at the same time as their parents were not backfilled for strictness. The effect was that
adding a new non-empty struct at the appropriate API level was
categorized as an ABI breaking change even though it was compatible
(the tool was confused into thinking the struct/member is being added
to an existing struct).

Also fixes a bug in fidl_api_compatibility_test.py where the user
was asked to check their @available annotation usage for acceptable
changes to the in-development API level.

Change-Id: Id4265a0cfee1f244a79e4fe47cf36bedcd9b12f0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/728302
Reviewed-by: Mitchell Kember <mkember@google.com>
Commit-Queue: Kendal Harland <kjharland@google.com>