[omaha-client] Update tests per RFC6761
RFC6761 defines special-use domain names, including domains to be used
in documentation. 'omaha-client' unit tests encode fuchsia.com,
foo.com, and bar.com, which are names that may exist in DNS. It can be
unclear whether reliance on the function of that domain within DNS is
relevant to the test.

By complying with the special-use domain names in these tests, there is
less cause for DNS concerns in this component.

Test: omaha-client

Fixed: 109427
Change-Id: I5264838e63f0c16f1012a9e7d992b639d21bc3a1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/725375
Commit-Queue: Antonio Linhart <antoniolinhart@google.com>
Reviewed-by: Sen Jiang <senj@google.com>