[wlan][policy] cancel autoconnect on new connect
Cancels any attempts to auto-connect to a network when there is a new
connect request. In some tests, a connect and disconnect could occur
while network selection is happening for auto-reconnect. In real life
this is highly unlikely to happen because the scans for the auto-connect
and manual connect couldn't happen at the same time. And the
auto-connect wouldn't replace and completed connection because it
will only connect to an idle iface.

Test: added unit test (fx test wlancfg-tests)
Bug: 108037
Change-Id: I9e8bf367b74126b16f08028a8b06e5dd12267f62
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/724464
Commit-Queue: Naomi Mccracken <nmccracken@google.com>
Reviewed-by: Kevin Sakuma <sakuma@google.com>