[ffx][fuzz] Add arguments for tests.json
This CL adds a '--json-file' argument to the `shell` and `list`
commands. This argument can be used to specify the location of
tests.json. In its absence, the code will still attempt to find
tests.json from FUCHSIA_DIR as before, but only when running one of
those two commands.

Change-Id: I00681db9cb10c4d41c9c063851667211c71e4a25
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/728166
Commit-Queue: Aaron Green <aarongreen@google.com>
Reviewed-by: Cameron Finucane <eep@google.com>