[dhcpv6] Process Reply to Renew
Process reply message received in response to a renew message, per RFC
8415 section 18.2.10.1
https://datatracker.ietf.org/doc/html/rfc8415#section-18.2.10.1.

Transition to AddressAssigned state on successfully renewing addresses.
Stay in Renewing and rely or retransmission timeout if an IA (regardless
if it is Assigned or ToRequest) is missing in the Reply.

Note: the No Binding case is handled in a follow up CL.

Test: fx test //src/connectivity/network/dhcpv6
Bug: 96319
Fixed: 96321
Change-Id: Ic494fc5e1baa487dcb5e21c1aba942e6ff00ca8c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/684349
Reviewed-by: Jake Fried <friedj@google.com>
Commit-Queue: Tony Gong <gongt@google.com>
Reviewed-by: Ghanan Gowripalan <ghanan@google.com>