[test_manager] Fix events routing in tests
Events routed to tests should be scoped to the test_root. There
shouldn't be both an event carrying the "test_root" name and an event
without it routed to tests.

Fixing this has the nice result of being able to delete the
diagnostics.rs moniker rewriting. However we can't do that until
the fix for the diagnostics CTF test lands and CTF releases catch up.

When the archivist receives a "." moniker it'll interpret it as "<root>"
for selector and presentation purposes.

Run-All-Tests: true
Bug: 81980

Change-Id: Ib3638a7d08df615c55f6650bf8f5e56f291e817f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/726942
Commit-Queue: Miguel Flores <miguelfrde@google.com>
Reviewed-by: Christopher Johnson <crjohns@google.com>
Reviewed-by: Brian Bosak <bbosak@google.com>