[netstack3] Give `do_send` public visibility
`do_send` was a `pub(crate)` function which depends on `TcpSyncContext`.
It needs to be `pub` because it will be used in a child CL, with change
I59922b0302bf27d77359ea5fc7a4301a3624419d, exporting this function needs
a method in `TcpSocketHandler`, which is done in this commit.

Test: No behavior change

Bug: 48578
Change-Id: I0ec6fcb9f7c64051d6471b233e1df02898c41d92
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/728670
Reviewed-by: Alex Konradi <akonradi@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Zeling Feng <zeling@google.com>