[driver_manager] Offer fuchsia.driver.compat to DFv2
When driver manager loads a DFv2 driver, give that driver
the fuchsia.driver.compat.Service if it has a DFv1 parent.

This lets the DFv2 child access
- topological path
- metadata

The following are still unimplemented:
- DFv1 FIDL offers
- Banjo

The lifecycle test is updated so the driver connects with
fuchsia.driver.compat.Service to get the topological
path correctly.

TEST: dfv2_v1_lifecyle

Change-Id: I2d315f5ebdb305a55511ced5034500ce74834fd9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/726323
Reviewed-by: Sarah Chan <spqchan@google.com>
Commit-Queue: David Gilhooley <dgilhooley@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>