[nxpfmac] Repurpose MoalContext into DeviceContext
With functionality being placed in different components the code base
has ended up in a place where a lot of pointers are being passed around
to access the most commonly used objects such as the event handler and
ioctl adapter.

Repurpose the MoalContext to be a DeviceContext to hold these pointers.
This context can then be passed around instead of a whole bunch of
pointers. Use the context for WlanInterface and Scanner construction
initially.

Test: Compiles, tests run
Change-Id: I5f07605a45a0eb1a443e70d10ae0825b1e45188a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/720265
Commit-Queue: Bjoern Johansson <bjoernj@google.com>
Reviewed-by: Karthik Krishnaswami <karthikrish@google.com>