[fastboot] Add delay before reboot
Some times the reboot is so fast that the devices hasn't been able to
sent the response packet over the transport. The consequence is that
host side fastboot hangs waiting for the response. Add a small delay as
a mitigation to make sure the response packet is sent.

Bug: b/217597389
Change-Id: I61ce98201c19f36681336a40bd4f4d48a1fef586
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/728862
Commit-Queue: Yecheng Zhao <zyecheng@google.com>
Reviewed-by: David Pursell <dpursell@google.com>