[audio] Add UnreadableClock
This CL adds UnreadableClock, which is just a `std::shared_ptr<Clock>`
where the only exposed method is `Clock::koid`.

PipelineStage and GainControl used to represent clocks using zx_koid_t,
but now use UnreadableClock. This change simplifies reference counting.
Before this CL, we could not determine if a clock is in use without
checking the koid of each PipelineStage and GainControl. After this CL,
we can check `std::shared_ptr<Clock>::use_count()` to determine if a
clock is still in use. A future CL will rely on this property.

Bug: 87651
Change-Id: Ic0adc4b68d7ff8a0a2d733a9474571659f1c8f18
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/728840
Reviewed-by: Alper Gungormusler <alperg@google.com>
Commit-Queue: Tom Bergan <tombergan@google.com>